Add google login to fronend. Refactor login to components
Add google auth to backend
