+++ backend/.env.example
@@ -1,0 +1,11 @@
+ATLAS_URI=
+GOOGLE_AUTH_CALLBACK=/auth/google/callback
+GOOGLE_CLIENT_ID=
+GOOGLE_CLIENT_SECRET=
+BASE_URL=http://localhost:5000
+CLIENT_CALLBACK_URL=http://localhost:3000/auth/google/success
+PORT=5000
+COOKIE_KEY =
+MONGO_URL=
+GITHUB_CLIENT_ID=
+GITHUB_CLIENT_SECRET=
