+++ client/src/pages/Login/index.scss
@@ -74,6 +74,11 @@
 .facebook {
   background-color: #507cc0;
 }
+
+.google {
+  background-color: darkgray;
+}
+
 input {
   width: 200px;
   padding: 15px 20px;
