+++ client/src/storage/ImageStorage.js
@@ -1,0 +1,11 @@
+import Google from "../assets/img/google.png";
+import Github from "../assets/img/github.png";
+import Facebook from "../assets/img/facebook.png";
+
+const imageStorage = {
+    google: Google,
+    github: Github,
+    facebook: Facebook
+}
+
+export default imageStorage;
